Add farming mechanics with Bushcraftable Planters
This commit adds new planters to the bushcrafting system. These can be
used to plant, grow and harvest seedlings. Simply craft the planter,
place it on the ground, drop a valid seedling ingredient onto it, and
water it regularly by dropping water vessels onto it. Once fully grown,
activate it to open the menu to harvest.

Planter/farming mechanics are compatible (and designed to work with)
graphic herbalism. When GH is installed, plants that have meshes that
still have bits of plant left after harvesting are deemed "recoverable",
and won't be destroyed when harvested. Instead, they will become
harvestable after a day or so.

Other changes include:
- Continue dehardcoding campfire code into Activator module
- Restore bounding boxes to a number of bushcrafted meshes
- Various bug fixes
- Use new addItem "showMessage" param